requirements.txt
streamlit 
groq 
python-dotenv
